demos/rag/requirements.txt
notion-client
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken

